+Kill process bugfix

+++ js/modules/rpc.interface.js
@@ -4,7 +4,7 @@
  * Time: 8:48 PM
  * To change this template use File | Settings | File Templates.
  */
-const DEV = true;
+const DEV = false;
 var settings = null;
 const http = require('http');
 const exec = require('child_process').exec;
@@ -28,8 +28,6 @@ if(os.platform() == 'win32'){
 }
 
 const _defaultRpcHeaders = {"jsonrpc":"2.0", "id":"0"};
-var currentSettingscurrentSettings;
-
 
 function _getRandomPort() {
     var charSet = '0123456789';
@@ -157,12 +155,12 @@ function init(appFolder, incommingSettings, onReady, onNetworkSync, onFail) {
 function destroy(){
     stopHalykcoinDeamon(function(){
         console.log("Halykcoin deamon is shouted down");
-        rpcServer.kill('SIGKILL');
+        if(rpcServer != null) rpcServer.kill('SIGKILL');
         process.exit();
     }, function(){
         console.log("Can't stop the deamon in normal way. Trying to kill the process");
-        rpcServer.kill('SIGKILL');
-        blockchainServer.kill('SIGKILL');
+        if(rpcServer != null) rpcServer.kill('SIGKILL');
+        if(blockchainServer != null) blockchainServer.kill('SIGKILL');
         process.exit();
     });
 }
